Have a considerable amount of fun bootstrapping the outline of a rust project.
I'm still learning how all the module stuff interacts, to be honest.
(And really everything about rust.)
But I think this is a pretty good starting point?

Includes a workspace with two crates within it.
(Partially this is because I know if one wants to develop proc macros,
one ends up needing a separate crate for that.  Otherwise it might be
overkill... but that future hurdle makes me want to do it immediately.)

Exercises a little bit of serde, and the clap CLI library.
And brings in a crate for order preserving maps
(which I'm thrilled to find serde also already knows of).

Includes some hello world testing of fixtures round tripping.

No serious features exist here.  But getting all the foundational setup
figured out took me long enough that I think a checkpoint is due :)