Macro time.
I consider it important to be able to have densely packed string
representations for some values when API building.

Serde doesn't seem to have *quite* the built-ins I wish for...
but there's some pieces laying around.

Breaking the problem down into "let's use Display and FromStr"
seems like a reasonable intermediate point.
Then there's macros in the serde_with crate that can cross the
rest of the gap nicely.

(There's some detritus in this checkpoint commit from before
I discovered that route.  It's more complicated, and I didn't
finish it before discovering this nicer route.  It's a monument
for now, and I'll remove it at the end of this journey.)

So... now we're generating Display and FromStr via macros.
I've been very surprised to find this doesn't seem to be done by
any existing crates I can find.  Or at least, not done very well.
I've found several crates that do it for enums, but only if they
have no values inside their members (e.g., if they're plain enums and
not sum types, then there's crates for it.  But I want sum types!).

And this is going... okay.  I'm learning how rust macros work as I go,
but I think this is converging on good stuff surprisingly rapidly.

In this checkpoint, we've got the Display implementation already.
The FromStr is still to come.  And a great deal of customization
is also still to come.  (I think this involves other macros, to
attach "attributes"?  I haven't fully learned that part yet.)

Fun!