Macro checkpoint.
Outline of FromStr impls.  I think it's about right for enums;
it's not at all started yet for structs.

Struct don't seem like they should be especially hard after this.
I just needed a bit of a checkpoint, because this... took a while.

There's some dangling code for trying to parse more annotations,
but it's not at all finished yet.  It's essentially research notes
at the moment, but slightly too far along to toss.

The hardest part of this, truthfully, was probably figuring out
what the heck is going on with error types.
(I'm admittedly a Rust novice.  I hope some of that dyn boxing
and the tons of implied conversion going on there and the weird extra
type syntaxes all make sense to me someday.  Today, they do not.)

This is, by the way, auto-formatted, as far as I can tell.
I'm using vscode in a pretty standard setup and rustfmt, and it
is modifying the buffer when I ask it to, so it's not noop'ing.
I don't understand why it's such a mess despite that.
Not a mystery I'm diving into right now.