Checkoint (broken) of trying to imrpove our error values.
The reason this is broken is because of another rule about crates
containing macros, which I just learned of: we can't export anything
ELSE.  Just macros.  (I knew one can't use a macro from within one's
own crate, but today I learned the rule is much bigger than that.)

So we can't export the error type that I'd now like to declare.

I don't think it's really possible to make this work without exporting
the error, or at least a constructor for it, either.  (And even if
it was possible, I don't think it would be a good idea.)

So... next commit up, we're gonna have ANOTHER crate in this workspace.

The current code will all move to a "catverters_derive" crate, which
I understand is the typical name Rustaceans choose for such things.
And we'll have a new "catverters" crate which is... dang near empty;
really, it's just going to contain... this error type.

Ah well.  As much as I lament some of these limitations (and mostly,
just being surprised by them), the overall power of this macro system
is so valuable that I really can't complain.