Split catverters and catverters-derive.  Works.
The catverters crate is doing nothing but holding a dang error type.

But, it's a really nice error type.  It prints good mesages!
Good error messages are important.

Git doesn't seem to recognize these as file moves, unfortunately, but
the vast majority of this diff is indeed just a move.

I have one nit with the code; I can't seem to get the more specific
error type to be part of my FromStr trait defn, and I can't quite
figure out why.  Would be nice to make it more specific.