One more form of ownership dance attempt, but still no dice.
I give up, after this.

My mental model is that the Vec is the owner of some underlying slice,
and as long as I'm "move"ing the Vec, I should be keeping ownership.
That should be sufficient for me to legally grab slices into the thing
that I (continuously) own.

I can't seem to express that.  The initial `as_slice` to peek into
the Vec is considered a borrow, and I can't figure out how to give
that any lifetime hints, and so overall I'm stuck.

I... must have some fundamental misunderstanding of memory ownership.
I really didn't expect this to be especially hard.