Salt the box with constraints.  Send and Sync seem somewhat capricious for Clap to require, IMO, but also apparently it's not a costly buck to pass, either.
This now appeases Clap.  (Okay, there are other compile errors in this
commit yet, but they will disappear in the next one.)

About whether it's bizarre to require Send and Sync traits of an
arbitrary error we might want to wrap:

- I'm certainly scared that this might be weird;
- it's fine in practice... so far...
- (but the only real inhabitant of that box right now is `Infallable`,
   which is certainly a fairly trivial type!  There's no reason to
   expect things that work for it will work in general)
- but also the reason that it works without fuss appaars to be
   "auto trait implementation".

This last bit is not something I'm super familiar with yet, still being
very much a rust novice, but *apparently*, it *seems*... this means that
Send and Sync will happen to be implemented on most types.

Anyway, I'm gonna roll with this and wait for it to bite me.
Maybe it never will.  If it does: it'll be a learning experience!  :D