Finish rigging sub-subcommands.  Finally compiles again.
One must make a new command struct for each subcommand before
going on to add another enum of subcommands.

Makes sense when you think about it.

(I don't happen to be introducing many flags on the middle steps of the
command tree... but, one _could_, so this arrangement has reason.)