YEEEAAAAHHHHH!  Real CLI wired now.  And **great errors**.
Try this:

`warpforge-cli catalog read-item heyo`

And you'll receive:

  > error: invalid value 'heyo' for '<CATALOG_REF>': failed to parse CatalogRef value: "heyo" should have more separators (the next one would delimit the start of the release_name field; the separator is ":")

Y E S.

It's hard to say how exciting this is.  It shouldn't be.  But it is.
Trying to get that quality level out of the golang libraries where
I've been spending time with for the past few years... has been an
exercise of pulling teeth, to put it mildly.  Here, by the time things
compile, it's all sorted.  The libraries are handling errors well and
attaching lots of contextual details (yay); and for the parts of the
errors that were my responsibility, we saw the compiler helping
shepherd things into a well-typed direction; and the combo of that plus
macro libraries that help write error messages... All of this worked
together to move things in a good direction.

I've tried to solve that second problem in golang by producing the
go-serum-analyzer (but it's pulling teeth, lots of work, has many
limitations, and is stuck as non-core influence levels), and I've tried
to solve that second problem with go-serum templating helpers (and
that's had some good success, but still requires a lot of boilerplate
to actually use).  By contrast: Rust's ecosystem just _has_ these
things already.  And they're _really good_.

Yes, this commit message is an ode to rust again.  Sorry not sorry.