Stretching my wings on traits vs impls, dynamic disbatch in practice, and some filesystem work.  Including some clever use of AsRef.
Probably none of this is interesting to a skilled rustacean,
but a bunch of it is an adventure when learning it the first time,
so this is a very baby-steps commit.

Mild gripe: PathBuf as the owned sibling of Path is... structurally, great,
but gosh I wish there was a standard naming convention or some other form
of discoverability for that pattern which is more automated than reading
the full prose docs and hoping they give you the key nudge about it.