Switch to using a proper type for the parameter here, and... I don't love the amount of syntax noise that generated.
It's a stylistic choice to use that many :: and package qualifiers;
that's not the issue.

The `.to_string()` to get an owned string... that does not bring joy.

Neither does the need to put an `&` in the callsite of load_module.
I thought that was supposed to be inferrable?

(The `&` on `module_name.0` makes perfect sense and I expect it there;
we *should* need to say that's a borrow and not move.)

I think these are things that are solvable by implementing more traits,
but gosh.  Which ones?  And: am I going to have to be constantly doing
this for *everything* in order to have nice usage?

Or is it that I need to do the trait bounds of AsRef trick constantly?
That's almost even more obnoxious, because for all that those are neat
and smooth when you see their usage site, they're absolutely wicked in
terms of producing docs: they don't give great clues for what concrete
types can be routed through the rule, at least to my reading skill
level so far.  And they about double the size of the type declaration
of a function.  And have two textual change-sites, at that.  So... I
*hope* I'm not going to feel inclined to that constantly.

(Also, added `pub` on the field of the ModuleName wrapper type struct,
and that's probably temporary assuming we actually want to do
validation with that type.)