okay, here's the one remaining diff cargo fmt makes for me.
And it's a dumb line-length driven one.
It makes that line (or now 3 lines) visually consistent with all the neighbors.
And it saves no actual length at all.
It's objectively stupid.

But okay.  I guess I can't win with this tooling as it stands today.

Biggest, biggest sigh.

I'm not mad, you're mad.