okay, rustfmt, this is a CRIMINALLY idiotic diff, and I'm really honestly angry.
- it makes the longest lines LONGER

- it's reflowing entire function call positions across lines -- that
  is NOT an acceptable behavior from a formatter.  That's *way* too
  much failure to preserve original user intent.

- and it just so happens to be hiding error handling and major
  flowcontrol by shoving it to the end of a long line.
  This is where I cross the line into calling it outright criminal:
  hidding my `map_err(...)?` from me is absolute
  Fuck OFF fuck OFF fuck ENTIRELY off. Fuck no.

And yet I have not found the knob to disable this; and I've tried --
see earlier commits today.

So I give up for now.

But I'm making this commit to document my extreme discontent.

...

I'm almost willing to change completely over to huge match statements
using explicit return on error arm and implicit return out of the match
on the okay arm.  That's a large amount of boilerplate, but at least
it leaves it very clear that there's an exit branching happening.
Which is otherwise, apparently, impossible to do within rustfmt's
absolutely bonkersly strong opinions.