non-compiling checkpoint in attempting to improve error handling of prototype gvisor executor.
I'm tumbling down a rabbithole here -- what I'm trying to get is
behavior indistinguishable from doing the serialization and the IO
in separate calls with my error handling control flow between them.
(I could very well just do that -- the only thing it would give up
is some very arguable concept of "streaming", which I most definitely
do not need to care about in this context whatsoever.)

I feel like I _should_ be able to do it, what with all of Rust's
delightful strong types and match statements and so on.

And I'm pretty puzzled.

I think serde_json::Error is missing some accessors that I want,
first of all.

I'm also stumped on an ownership thing.  That's the bit that's not
compiling right now.  I can't seem to figure out how to take enough
ownership of the stuff I'm unwrapping from inside the serde_json::Error
in order to be allowed to return it, and I'm... not sure why.