additional flailing attempts to own error values enough to manipulate them.
Salting a move on it didn't help.

Decomposed the chain of calls line by line, for visibility.

Made sure to get rid of later references to 'e',
although that was not relevant.  (Good to confirm, at least.)

Latest try was an `.into()` but that was kinda a shot in the dark,
and, nope.