check out the expect-test crate.  It's... good.
The ability to regenerate test fixtures is ultraexcellent.
This crate does that pretty nicely.

I haven't tested its capabilities with external files yet,
but apparently that's also supported.

One nit I have is that when it regenerates things,
it utterly ignores the previous leading whitespace.
That seems... silly, but also quite fixable.
I might try to PR that into the upstream in the future.
Similarly, you can't really have a trailing linebreak in order to
put all the closing bojangles -- literally, for this example
,all of `"#]])` -- four characters of noise! -- on a separate line.
But this is a very minor complaint indeed.

I rewrote what this test case is doing quite a bit to make it
use pretty print.  It now defacto demands the fixture be prettyprinted.
That wasn't the original point of the test, but it does make it easier
to maintain and regenerate, so, fair enough.
Bit of a shame how much boilerplate that was in serde_json.  Didn't
expect that.  (A two-space indented prettyprint is still a oneliner;
the next most incremental step beyond that suddenly explodes to this
5 or 6 line territory.  Uff.  Oh well!)

Overall: works.  And regenerable fixtures: NICE.