So I did a spicy thing today, which is write a yaml serializer to quote all strings unless they come from enum or struct fields.
It kinda works.  You can see the tests.

Many things are still pretty wild about this.

- I don't think I brought over the plz-use-another-map-instead-of-yaml-tags behavior yet.  Didn't test.
- I didn't check of maps DWIM with strings yet.  If they don't, that's probably an easy fix.
- Wow that's this is a lot of code for such a minor outcome.
- It only does single quoting.  Not double.  Apparently the underlying library only supports double for parsing, but not for emitting.
  Weird.  Okay.  Changing that sounds like more work than I want today.
- Looking deeper at this whole thing... wow, there's... a lot of c/ffi/unsafe code in the bottom of libyml.  It does not bring joy.
  And boy does that ever redouble, or triple, or further... that I'm _not_ gonna try to get double quotes wormed in down there.

So I'm not sure where this leaves me aiming next.
This wasn't a particularly expensive experiment,
and it doesn't take us very far (if at all) outside yaml norms,
so it's... viable as-is.

But it's pretty alarming looking at the bottom of it.
I'm wondering if it would be better to pick a different library
that's more-actually-rust for the yaml on the bottom,
and then do the serde bindings on top of that (since the current
rust ecosystem seems to be in limbo and lack any other one of those
except this one seen here), since the serde binding effort is seemingly
not super huge.