update cargo lock.
Originally, was just trying to solve merge conflicts, but it
seemed easier to just blow everything away and let it re-select.

Which of course selected different things.  Some newer.
derive_more in particular changed a bit.

I seem to have picked up a more aggressive autoformatter again
too, about which I'm not even going to ask.

Anyway, hopefully this is a zero-impact diff, in the end.