warpforge-cfg: introduce!  Config should be gathered somewhere.
And here are my initial thoughts on how we should start splitting
apart data storage locations, while taking some cues from the XDG
standards.

Previous iterations of warpforge work have gathered *everything*
in the workspace root, and... that works, but is not very satisfying.
(My personal experience with that was it left large amounts of
files scattered in various parts of the filesystem that either
required exceptions to be made to by backup and sync tools, or,
required configuring every dang workspace to store things elsewhere.
So, this time, let's try some different defaults that should be
less likely to provoke those problems, and take ecosystemically
common cues for where people want things like "caches" to be.)

References to "depot" here are still forward looking.
That's up next.