warpforge-depot: introduce, starting with readme outlining the scope.
This is addressing a series of problems that we previously tried to solve
in a project called "rio" (short for "repeatable IO"), but...

- again
- with a better name
- in rust
- and paying much more attention to the most common case, where most
  attributes *don't* matter, and being able to run without advanced
  privileges is by far the prevailing and overriding concern.

(That last one is a pretty high impact detail.)

I think the scope *might* still involve exec-based plugins,
but we may also end up simply porting a good fraction of what
rio used to do into rust, because it'll be easier than maintaining
two code bases.  We'll see.